requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--index-url http://localhost:3141/root/pypi/+simple/

django-parler==1.9.2
django==1.11.16
pytz==2018.7              # via django
